Move agent roles to 'agent.roles' module
