+++ src/agents/__init__.py
@@ -3,10 +3,10 @@
 from .agent import Agent
 from .base import BaseAgent, BaseWriterAgent
 
-from .planner import Planner
-from .internal_monologue import InternalMonologue
-from .researcher import Researcher
-from .formatter import Formatter
-from .coder import Coder
-from .action import Action
-from .runner import Runner
+from .roles.planner import Planner
+from .roles.internal_monologue import InternalMonologue
+from .roles.researcher import Researcher
+from .roles.formatter import Formatter
+from .roles.coder import Coder
+from .roles.action import Action
+from .roles.runner import Runner
