+++ src/agents/roles/decision/prompt.jinja2
