+++ src/agents/roles/formatter/formatter.py
