+++ src/agents/roles/formatter/prompt.jinja2
