+++ src/agents/roles/internal_monologue/__init__.py
