+++ src/agents/roles/internal_monologue/internal_monologue.py
