+++ src/agents/roles/internal_monologue/prompt.jinja2
