+++ src/agents/roles/researcher/prompt.jinja2
