+++ src/agents/roles/researcher/researcher.py
